add and use custom permissions classes

+++ api/views/question_views.py
@@ -3,6 +3,7 @@ from rest_framework.views import APIView
 from api.models import *
 from django.shortcuts import get_object_or_404
 from rest_framework import status
+from rest_framework.exceptions import PermissionDenied
 from api.serializers import (
     QuestionMultipleChoiceSerializer,
 )
@@ -10,13 +11,18 @@ from rest_framework import serializers
 from django.http import JsonResponse
 from drf_spectacular.utils import extend_schema, OpenApiResponse
 
+from ..permissions import IsQuestionOwner
+
 
 class QuestionView(APIView):
+    permission_classes = [IsQuestionOwner]
+
     @extend_schema(
         request=QuestionMultipleChoiceSerializer(many=True),
         responses={
             201: OpenApiResponse(description="Questions created successfully"),
             400: OpenApiResponse(description="Bad Request"),
+            403: OpenApiResponse(description="Forbidden"),
         },
     )
     def post(self, request):
@@ -35,8 +41,12 @@ class QuestionView(APIView):
                 try:
                     serializer = QuestionMultipleChoiceSerializer(data=question_data)
                     serializer.is_valid(raise_exception=True)
+                    quiz = Quiz.objects.get(id=question_data["quiz_id"])
+                    self.check_object_permissions(request, quiz)
                     serializer.save()
                     created_questions.append(serializer.data)
+                except PermissionDenied as e:
+                    return JsonResponse({"detail": e.__str__()}, status=status.HTTP_403_FORBIDDEN)
                 except serializers.ValidationError as e:
                     errors.append({"question_data": question_data, "error": e.detail})
                 except Exception as e:
@@ -62,6 +72,7 @@ class QuestionView(APIView):
     # Put method to update text of a question by taking in an id
     def put(self, request, question_id):
         question = get_object_or_404(QuestionMultipleChoice, id=question_id)
+        self.check_object_permissions(request, question)
         question.__dict__.update(request.data)
         question.save()
         return JsonResponse({"message": "Question updated successfully"})
@@ -69,6 +80,7 @@ class QuestionView(APIView):
     # delete method to delete question by id
     def delete(self, request, question_id):
         question = get_object_or_404(QuestionMultipleChoice, id=question_id)
+        self.check_object_permissions(request, question)
         question.delete()
         return JsonResponse({"message": "Question deleted successfully"})
